requirements.txt
cffi==1.14.1
chardet==3.0.4
click==7.1.2
cryptography==3.0
dnspython==2.0.0
email-validator==1.1.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.1
pdfminer.six==20200726
pkg-resources==0.0.0
pycparser==2.20
PyMySQL==0.10.0
python-slugify==4.0.1
six==1.15.0
sortedcontainers==2.2.2
SQLAlchemy==1.3.18
text-unidecode==1.3
Werkzeug==1.0.1
WTForms==2.3.2
